electron/backend/requirements.txt
uvicorn[standard]
fastapi
PyYAML
faiss-cpu
llama-cpp-python
sentence-transformers